requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
python-dotenv==1.0.1
pytest==8.3.5
httpx==0.27.2
requests==2.32.3